+++ dp/__init__.py
@@ -1,6 +1,8 @@
 """Main library"""
 from dp._dp_array import DPArray
+from dp._logger import Logger
 
 __all__ = [
     "DPArray",
+    "Logger",
 ]