+++ HISTORY.md
@@ -2,5 +2,6 @@
 
 ## 0.0.0
 
+- Create ``Logger`` class (#4)
 - Create ``DPArray`` class (#3)
 - Initial project setup (#1, #2)